Pull out plain rules file loading code into a new `agent_rules` crate
