+++ crates/agent/Cargo.toml
@@ -19,6 +19,7 @@ test-support = [
 ]
 
 [dependencies]
+agent_rules.workspace = true
 anyhow.workspace = true
 assistant_context_editor.workspace = true
 assistant_settings.workspace = true
