+++ crates/agent_rules/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "agent_rules"
+version = "0.1.0"
+edition.workspace = true
+publish.workspace = true
+license = "GPL-3.0-or-later"
+
+[lints]
+workspace = true
+
+[lib]
+path = "src/agent_rules.rs"
+doctest = false
+
+[dependencies]
+anyhow.workspace = true
+fs.workspace = true
+gpui.workspace = true
+prompt_store.workspace = true
+util.workspace = true
+worktree.workspace = true
+
+[dev-dependencies]
+indoc.workspace = true
